backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
rembg==2.0.50
Pillow==10.3.0
aiofiles==23.2.1
python-dotenv==1.0.1
